Use provider for home scope
